adjust non-predictions slightly
use a positive large number (999) instead of a negative (-999).
needed for unichill since it uses doy_estimator as an mid-step
for calculating chill days required for head accumulation to
being.
have to adjust the unichill known values slightly for this, but now
it should be "correct". No other known values needed correcting.