redo structure of input data
with most models the only predictor variable is temperature,
so that was the sole argument besides observations. Newer
models require things like lat/long, daylength, precip, etc.

This replaces all prior temperature calls with "predictors",
where temperature can be one of many aspects. Required
predictors are defined in the model. See the M1 model
for an example.
Since most models still just use temperature, that default
is defined in the base model with some methods to check for it.
If a model requies something besides temperature it must be
defined within the model.

Added a new model, M1, which uses a new variable, daylength.
Also updated the example data to include lat/long and daylength