deal with leap year issues in temperature data
when you have temperature data spanning leap years and
non-leap years there can be a mismatch. For example, if
the first date temperature is Oct. 1, that is doy -92 on
a non-leap year but doy -93 on a leap year. In the dataframe
where every column is a doy, this creates "missing" entries
for doy -93. This was hard coded to drop doy -67 because
thats how all the example data was setup. But now it's
more robust in just checking the first day and issuing a
warning if it drops it.