propagate aggregation for ensemble prediction
If I make an ensemble of ensembles the aggregation type
would not be passed to the sub-ensebmels.

ie. An ensemble of 4 models, each of which is a bootstrap of
100 models, using aggregation='none' would return the 4
values, each of which is the mean of the bootstraps.

need **kwargs in base prediction method so it accepts, but
doesn't use, the aggregation arg.