update randomness in tests (#119)
expected values are constantly changing slightly with new python and scipy/numpy versions. I *think* the main reason is underlying randomness is changing. The newish numpy randomness ring is supposed to be more reproducable, so use that here and update all the expected values to see if it fixes things. Will only know after a few months to a year of underlying package updates.
Also this is keeping the expected values tests as just warnings if they fail as I don't think they are showstoppers.